Delete M6 model & Change M7 -> M6
