+++ README.md
@@ -29,11 +29,11 @@ If you want to test the M3 model, enter the following command in the shell:
 ```
 python main.py 3
 ```
-You can enter values from 1 to 7.
+You can enter values from 1 to 6.
 
 ## Dataset
 
-We use MNIST, CIFAR10, and USPS datasets for 2D CNN models which can be downloaded using Torchvision. The details about datasets can be found at https://pytorch.org/vision/0.15/datasets.html.
+We use MNIST and CIFAR10 datasets for 2D CNN models which can be downloaded using Torchvision. The details about datasets can be found at https://pytorch.org/vision/0.15/datasets.html.
 
 ## HE Parameters
 We use the Cheon-Kim-Kim-Song(CKKS) scheme following parameter settings:
@@ -62,8 +62,7 @@ We conducted seven experiments to test the performance of UniHENN with various m
 |M3|2D CNN|MNIST|7|M3_test.py|
 |M4|2D CNN|MNIST|12|M4_test.py|
 |M5|2D CNN|Cifar-10|11|M5_test.py|
-|M6|2D CNN|USPS|6|M6_test.py|
-|M7|1D CNN|ECG Dataset|7|M7_test.py|
+|M6|1D CNN|ECG Dataset|7|M6_test.py|
 
 ## Contributing
 - Main Contributor: Jihun Kim (https://github.com/JihunSKKU), Hyunmin Choi(https://github.com/hm-choi)
